Added Appointment-, Customer-, CustomerUpload- and PartService + some…

+++ src/main/java/nl/carsforyou/garage/repositories/PartRepository.java
@@ -4,5 +4,5 @@ package nl.carsforyou.garage.repositories;
 import nl.carsforyou.garage.entities.PartEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartRepository extends JpaRepository<PartEntity, Integer> {
+public interface PartRepository extends JpaRepository<PartEntity, Long> {
 }
